backend/requirements.txt
django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
channels>=4.0
daphne>=4.1
psycopg2-binary>=2.9
dj-database-url>=2.0
whitenoise>=6.6
gunicorn>=21.0
